# lib/api-zod/src/generated/types/bet.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zombonk Gambling API
 * OpenAPI spec version: 0.1.0
 */
import type { BetDetails } from './betDetails';

export interface Bet {
  id: number;
  playerId: number;
  gameId: number;
  gameTitle?: string;
  gameType?: string;
  wager: number;
  payout: number;
  won: boolean;
  details?: BetDetails;
  createdAt: string;
}