# lib/api-zod/src/generated/types/betActivity.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zombonk Gambling API
 * OpenAPI spec version: 0.1.0
 */

export interface BetActivity {
  id: number;
  playerName: string;
  gameTitle: string;
  gameType: string;
  wager: number;
  payout: number;
  won: boolean;
  createdAt: string;
}