# lib/api-zod/src/generated/types/gameOptionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zombonk Gambling API
 * OpenAPI spec version: 0.1.0
 */

export interface GameOptionInput {
  label: string;
  odds: number;
  emoji?: string;
  weight?: number;
}